リファクタリング

+++ src/main/java/databullet/domain/definition/dataspec/DataSpecTable.java
@@ -10,14 +10,14 @@ import java.util.List;
 @NoArgsConstructor
 public class DataSpecTable {
     private String name;
-    private Long rowCount;
+    private Integer rowCount;
     private List<DataSpecColumn> columns = new ArrayList<>();
 
     public static DataSpecTable empty(String name) {
         DataSpecTable table = new DataSpecTable();
         table.setName(name);
         table.setColumns(new ArrayList<>());
-        table.setRowCount(0L);
+        table.setRowCount(0);
         return table;
     }
 }
